caption-generation/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore
google-cloud-aiplatform
requests